+++ www/go/core/language/zh_tw.php
@@ -449,6 +449,7 @@ return array (
     'KZ' => 'Kazakhstan',
     'KE' => 'Kenya',
     'KI' => 'Kiribati',
+	'XK' => 'Kosovo',
     'KW' => 'Kuwait',
     'KG' => 'Kyrgyzstan',
     'LA' => 'Lao People\'s Democratic Republic',