+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - LDAP: Disable creation of IMAP accounts didn't work
 - Newsletter: Bug when an image was both attached as attachment and inline.
 - ActiveSync: duplicate Reference and in-reply-to header
+- Core: Added Kosovo to the countries list
 
 09-11-2022 6.6.128
 - Core: rollback of fix for autocomplete. Probably fixed by always creating an address book entry